Move static logic to backend/static.go, fix JSON content type
